refactor: refactor theme
