+++ packages/iceworks-client/src/appConfig.js
@@ -2,8 +2,20 @@ export default window.iceworksConfig || {};
 
 const PLACEHOLDER_IMG = '//img.alicdn.com/tfs/TB1lCcKc8Kw3KVjSZFOXXarDVXa-350-200.png';
 const THEMES = {
-  dark: '@alifd/theme-iceworks-dark',
-  light: '@alifd/theme-iceworks-light',
+  dark: {
+    themePackage: '@alifd/theme-iceworks-dark',
+    termTheme: {
+      foreground: '#fff',
+      background: '#333646',
+    },
+  },
+  light: {
+    themePackage: '@alifd/theme-iceworks-light',
+    termTheme: {
+      foreground: '#333',
+      background: '#f7f8fa',
+    },
+  },
 };
 
 export {
