+++ packages/iceworks-client/src/pages/Setting/components/General/index.js
@@ -5,7 +5,6 @@ import { Grid, Radio } from '@alifd/next';
 import { injectIntl, FormattedMessage } from 'react-intl';
 import { LocalContext, localeInfos } from '@components/Locale';
 import { ThemeContext } from '@components/ThemeProvider';
-import useTermTheme from '@hooks/useTermTheme';
 import Card from '@components/Card';
 import { THEMES } from '@src/appConfig';
 import socket from '@src/socket';
@@ -19,7 +18,6 @@ const General = ({ intl }) => {
   const { locale, setLocale } = useContext(LocalContext);
   const { theme, setTheme } = useContext(ThemeContext);
   const [editor, setEditor] = useState(DEFAULT_EDITOR);
-  const { setTermTheme } = useTermTheme(theme);
 
   // language options
   const languageOptions = Object.keys(localeInfos).map(key => {
@@ -34,11 +32,11 @@ const General = ({ intl }) => {
   const themeOptions = [
     {
       label: intl.formatMessage({ id: 'iceworks.setting.general.theme.light' }),
-      value: THEMES.light,
+      value: THEMES.light.themePackage,
     },
     {
       label: intl.formatMessage({ id: 'iceworks.setting.general.theme.dark' }),
-      value: THEMES.dark,
+      value: THEMES.dark.themePackage,
     },
   ];
 
@@ -71,7 +69,6 @@ const General = ({ intl }) => {
   async function onThemeChange(currentTheme) {
     await socket.emit('home.setting.setTheme', { theme: currentTheme });
     setTheme(currentTheme);
-    setTermTheme(currentTheme);
   }
 
   async function getEditor() {
